refactor: organize common components
