+++ src/components/common/BadgeCarousel.vue
