+++ src/components/common/StepIndicator.vue
